Split up FunctionCall
